requirements.txt
python-telegram-bot==21.5
APScheduler==3.10.4
requests==2.32.3
pytz==2024.2
